Moving rule selection to new commonrules.select package #53
