--- querqy-core/src/main/java/querqy/rewrite/commonrules/model/FilterCriterion.java
@@ -1,7 +1,0 @@
-package querqy.rewrite.commonrules.model;
-
-public interface FilterCriterion {
-
-    boolean isValid(final Instructions instructions);
-
-}
