+++ querqy-core/src/main/java/querqy/rewrite/commonrules/select/FilterCriterion.java
@@ -1,0 +1,9 @@
+package querqy.rewrite.commonrules.select;
+
+import querqy.rewrite.commonrules.model.Instructions;
+
+public interface FilterCriterion {
+
+    boolean isValid(final Instructions instructions);
+
+}
