+++ querqy-core/src/main/java/querqy/rewrite/commonrules/select/FlatTopRewritingActionCollector.java
@@ -1,6 +1,8 @@
-package querqy.rewrite.commonrules.model;
+package querqy.rewrite.commonrules.select;
 
 import querqy.PriorityComparator;
+import querqy.rewrite.commonrules.model.Action;
+import querqy.rewrite.commonrules.model.Instructions;
 
 import java.util.Comparator;
 import java.util.List;
