+++ querqy-core/src/main/java/querqy/rewrite/commonrules/select/RuleSelectionParams.java
@@ -1,4 +1,4 @@
-package querqy.rewrite.commonrules;
+package querqy.rewrite.commonrules.select;
 
 public interface RuleSelectionParams {
