+++ querqy-core/src/main/java/querqy/rewrite/commonrules/select/SelectionStrategy.java
@@ -1,4 +1,4 @@
-package querqy.rewrite.commonrules.model;
+package querqy.rewrite.commonrules.select;
 
 public interface SelectionStrategy {
