+++ querqy-core/src/main/java/querqy/rewrite/commonrules/select/Sorting.java
@@ -1,4 +1,6 @@
-package querqy.rewrite.commonrules.model;
+package querqy.rewrite.commonrules.select;
+
+import querqy.rewrite.commonrules.model.Instructions;
 
 import java.util.Collections;
 import java.util.Comparator;
