+++ querqy-core/src/main/java/querqy/rewrite/commonrules/select/TopLevelRewritingActionCollector.java
@@ -1,4 +1,8 @@
-package querqy.rewrite.commonrules.model;
+package querqy.rewrite.commonrules.select;
+
+import querqy.rewrite.commonrules.model.Action;
+import querqy.rewrite.commonrules.model.Instructions;
+import querqy.rewrite.commonrules.model.Limit;
 
 import java.util.Collection;
 import java.util.Collections;
