+++ querqy-core/src/main/java/querqy/rewrite/commonrules/select/TopRewritingActionCollector.java
@@ -1,4 +1,7 @@
-package querqy.rewrite.commonrules.model;
+package querqy.rewrite.commonrules.select;
+
+import querqy.rewrite.commonrules.model.Action;
+import querqy.rewrite.commonrules.model.Instructions;
 
 import java.util.List;
 import java.util.function.Function;
