+++ querqy-core/src/test/java/querqy/rewrite/commonrules/SimpleCommonRulesRewriterFactoryTest.java
@@ -16,7 +16,8 @@ import querqy.model.Term;
 import querqy.rewrite.QueryRewriter;
 import querqy.rewrite.SearchEngineRequestAdapter;
 import querqy.rewrite.commonrules.model.RulesCollection;
-import querqy.rewrite.commonrules.model.SelectionStrategy;
+import querqy.rewrite.commonrules.select.SelectionStrategy;
+import querqy.rewrite.commonrules.select.SelectionStrategyFactory;
 
 import java.io.IOException;
 import java.io.StringReader;
