+++ querqy-core/src/test/java/querqy/rewrite/commonrules/model/DeleteInstructionTest.java
@@ -1,9 +1,8 @@
 package querqy.rewrite.commonrules.model;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.contains;
 import static querqy.QuerqyMatchers.*;
-import static querqy.rewrite.commonrules.SelectionStrategyFactory.DEFAULT_SELECTION_STRATEGY;
+import static querqy.rewrite.commonrules.select.SelectionStrategyFactory.DEFAULT_SELECTION_STRATEGY;
 
 import java.util.Collections;
