+++ querqy-core/src/test/java/querqy/rewrite/commonrules/model/InputTest.java
@@ -6,7 +6,6 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-import org.junit.Before;
 import org.junit.Test;
 
 import querqy.ComparableCharSequence;
