+++ querqy-core/src/test/java/querqy/rewrite/contrib/ShingleRewriteTest.java
@@ -11,15 +11,13 @@ import querqy.rewrite.commonrules.AbstractCommonRulesTest;
 import querqy.rewrite.commonrules.CommonRulesRewriter;
 import querqy.rewrite.commonrules.LineParser;
 import querqy.rewrite.commonrules.model.Input;
-import querqy.rewrite.commonrules.model.Instructions;
-import querqy.rewrite.commonrules.model.InstructionsTestSupport;
 import querqy.rewrite.commonrules.model.RulesCollection;
 import querqy.rewrite.commonrules.model.RulesCollectionBuilder;
 import querqy.rewrite.commonrules.model.SynonymInstruction;
 import querqy.rewrite.commonrules.model.TrieMapRulesCollectionBuilder;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static querqy.QuerqyMatchers.*;
-import static querqy.rewrite.commonrules.SelectionStrategyFactory.DEFAULT_SELECTION_STRATEGY;
+import static querqy.rewrite.commonrules.select.SelectionStrategyFactory.DEFAULT_SELECTION_STRATEGY;
 import static querqy.rewrite.commonrules.model.InstructionsTestSupport.instructions;
 
 /**
