+++ querqy-for-lucene/querqy-solr/src/main/java/querqy/solr/PrimitiveValueSelectionStrategyFactory.java
@@ -2,8 +2,8 @@ package querqy.solr;
 
 import org.apache.lucene.analysis.util.ResourceLoader;
 import org.apache.solr.common.util.NamedList;
-import querqy.rewrite.commonrules.SelectionStrategyFactory;
-import querqy.rewrite.commonrules.model.SelectionStrategy;
+import querqy.rewrite.commonrules.select.SelectionStrategyFactory;
+import querqy.rewrite.commonrules.select.SelectionStrategy;
 
 public class PrimitiveValueSelectionStrategyFactory implements FactoryAdapter<SelectionStrategyFactory> {
 
@@ -11,7 +11,7 @@ public class PrimitiveValueSelectionStrategyFactory implements FactoryAdapter<Se
     public SelectionStrategyFactory createFactory(final String strategyId, final NamedList<?> args,
                                                   final ResourceLoader resourceLoader) {
         // TODO: pass strategyId
-        return new querqy.rewrite.commonrules.PrimitiveValueSelectionStrategyFactory();
+        return new querqy.rewrite.commonrules.select.PrimitiveValueSelectionStrategyFactory();
     }
 
     @Override
